+++ graph_layout/lib/graph_layout.dart
@@ -3,3 +3,4 @@ library graph_layout;
 export 'src/data_structures.dart';
 export 'src/spring.dart';
 export 'src/common.dart';
+export 'src/layout_algorithms.dart';