+++ graph_layout/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * Add basic data structures: `Node`, `IntegerNode`, `Edge` (undirected), `Graph`.
 * Add basic typedefs: `AdjacencyList`, `EdgeList`, `NodeLayout`.
+* Add abstract class `InteractiveLayoutAlgorithm`, which represents a generic interactive layout algorithm.
 * Add spring based layout algorithm `SpringSystem`, with support for constrained (fixed) nodes.